ajustes en factura legal y cliente

+++ src/main/java/com/franco/dev/graphql/productos/resolver/ProductoResolver.java
@@ -101,7 +101,7 @@ public class ProductoResolver implements GraphQLResolver<Producto> {
         List<Sucursal> sucursalList = sucursalService.findAll2();
         for (Sucursal s : sucursalList ){
             ExistenciaCostoPorSucursal eps = new ExistenciaCostoPorSucursal();
-            eps.setExistencia(movimientoStockService.stockByProductoIdAndSucursalId(p.getId(), s.getId()));
+            eps.setExistencia(movimientoStockService.stockByProductoIdAndSucursalId(p.getId(), s.getId()).floatValue());
             eps.setSucursal(s);
             CostoPorProducto cps = costosPorProductoService.findLastByProductoId(p.getId()).orElse(null);
             MovimientoStock ms = null;
